Remove local-cli
Summary:
This diff removes all files from the local-cli from react-native-github. The only leftover is the `cli.js` which redirects to `react-native-local-cli` and the templates folder which I'll move in a follow-up.

This diff simply removes all files, it makes no code changes. There will be a cleanup diff on top of this one.

Reviewed By: TheSavior

Differential Revision: D13338367

fbshipit-source-id: 1afabdb34202596fb13186d728df74482c9bc609