Add deprecated-react-native-listview
Summary:
This adds the deprecated ListView package to xplat/js. I published this module in https://github.com/facebookarchive/react-native-deprecated-modules

The way it is published I tried to reduce duplication of dependencies as much as possible. With this diff and the changes in package.json, the bundle size should not regress (until the actually deprected ListView is included; and then it will be recovered once ListView is fully removed from React Native, which will happen within a short timeframe). The change to `react-clone-referenced-element` are non-breaking.

Reviewed By: TheSavior

Differential Revision: D14163988

fbshipit-source-id: 0fc2d1f33855a8b1a709f5316e68c734716ec049